% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate_dropper.R
\name{associate_dropper}
\alias{associate_dropper}
\title{remove collection from associated species}
\usage{
associate_dropper(x, Binomial)
}
\arguments{
\item{x}{a data frame containing collection info}

\item{Binomial}{a column containing the name of the collection, without authorship.}
}
\description{
this function removes the collected species from the list of associated
species
}
\examples{
out <- associate_dropper(collection_examples, Binomial = 'Full_name') |>
 dplyr::select(Full_name, Associates)
}
