% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associates_spell_check.R
\name{associates_spell_check}
\alias{associates_spell_check}
\title{Format associated species, and spell check them internally}
\usage{
associates_spell_check(x, column, path)
}
\arguments{
\item{x}{a data frame containing collection info}

\item{column}{a column containing the family information}

\item{path}{a path to a folder containing the taxonomic data.}
}
\description{
Apply proper capitalization to the species, and spell check all components.
}
\examples{
\dontrun{
df <- data.frame(
 Family = c('Apiacace', 'Apiaceae', 'Ongraceae', 'Syperaceae', 'Poacaceae')
)
p2tax <- '/media/steppe/hdd/Barneby_Lives-dev/taxonomic_data'
associates_spell_check.R(x = df, column = 'Family', p2tax)
}
}
