% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associates_writer.R
\name{associates_writer}
\alias{associates_writer}
\title{do or don't write the associated plant species}
\usage{
associates_writer(x)
}
\arguments{
\item{x}{dataframe holding the values}
}
\description{
This functions determines whether to write information on associated plant
species. It used as a final step when formatting labels.
}
\examples{
library(BarnebyLives)
hilaria <- collection_examples[73,]
associates_writer(hilaria)
agoseris <- collection_examples[82,]
associates_writer(agoseris)
}
