% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofill_checker.R
\name{autofill_checker}
\alias{autofill_checker}
\title{Search for accidental auto-increments from spreadsheet autofills}
\usage{
autofill_checker(x)
}
\arguments{
\item{x}{a dataframe which has undergone 'dms2dd' from BarnebyLives}
}
\description{
Spreadsheets like to automatically increment everything all the time.
Here we try to detect these where they may have effected the coordinates.
This function operates under the observation, that the incrementation only occurs
to the 'integers' or degrees
}
\examples{
coords <- data.frame(
  longitude_dd = c(rep(42.3456, times = 5), 43.3456),
  latitude_dd = c(rep(-116.7890, times = 5), -115.7890)
)
autofill_checker(coords) # note that all values in the column will
flagged after the occurrence (see 'Lat_AutoFill_Flag')
}
