% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autonym.R
\name{autonym}
\alias{autonym}
\title{retrieve author results for autonyms}
\usage{
autonym(genus, epithet, infrarank, infraspecies)
}
\arguments{
\item{genus}{derived from result_grabber}

\item{epithet}{derived from result_grabber}

\item{infrarank}{derived from result_grabber}

\item{infraspecies}{derived from result_grabber}

\item{export}{}
}
\description{
this function is internal to result_grabber, it ensures that an author is collected from the KEW API for infra-species which have autonyms.
}
