% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{check_data_setup_outputs}
\alias{check_data_setup_outputs}
\title{Check whether expected output files from data_setup() exist}
\usage{
check_data_setup_outputs(pathOut)
}
\arguments{
\item{pathOut}{Output directory used in data_setup()}
}
\value{
None. prints to console.
}
\description{
Check whether expected output files from data_setup() exist
}
\examples{
\dontrun{
check_data_setup_outputs(path.expand('~/Documents/BL_testing_data/geo'))
}
}
