% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_writer.R
\name{collection_writer}
\alias{collection_writer}
\title{format collector info and codes}
\usage{
collection_writer(x)
}
\arguments{
\item{x}{dataframe holding the values}
}
\description{
Format these data to have commas or periods as appropriate
}
\examples{
data('collection_examples')
ce <- collection_examples[14,]
collection_writer(ce)
}
