% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{data_setup}
\alias{data_setup}
\title{Set up the downloaded data for a BarnebyLives instance}
\usage{
data_setup(path, pathOut, bound, cleanup = FALSE)
}
\arguments{
\item{path}{Character. path to the folder where output from \code{data_download} was downloaded.
Defaults to a subfolder within the working directory ('./geodata_raw').}

\item{pathOut}{Character. Path to the folder where final spatial data should be placed.
Defaults to ('.'), and will create a 'geodata' directory within it to store results.}

\item{bound}{Data frame. 'x' and 'y' coordinates of the extent for which the BL instance will cover.}

\item{cleanup}{Boolean. Whether to remove the original zip files which were downloaded. Defaults to FALSE.
Either mode of running the function will delete the uncompressed zip files generated during the process.
If FALSE, this will also bundle the downloaded topographic variables into their own directories,
e.g. all directories of 'aspect' rasters will go into a single 'aspect' directory.}
}
\description{
Once all data are downloaded for BarnebyLives use this function
to set them up to be used by the pipeline.
}
\examples{
\dontrun{
bound <- data.frame(
  y = c(42, 42, 44, 44, 42),
  x = c(-117, -119, -119, -117, -117)
)

setwd('/media/steppe/hdd/BL_sandbox/geodata_raw')
data_setup(path = '.', pathOut = '../geodata', bound = bound, cleanup = FALSE)
}
}
