% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_templates.R
\docType{data}
\name{database_templates}
\alias{database_templates}
\title{database_templates}
\format{
data frame
}
\usage{
database_templates
}
\description{
Culled list of templates for different herbarium database management systems.
}
\details{
this data.frame is opporunistically gathered from a variety of herbaria. It shows the formats
which a spreadsheet document can be ordered to for mass uploads.
}
\examples{
head(database_templates)
}
\keyword{datasets}
