% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date2text.R
\name{date2text}
\alias{date2text}
\title{take mdy format date and make it written herbarium label format}
\usage{
date2text(x)
}
\arguments{
\item{x}{a data frame with dates}
}
\description{
take the mdy date format, popular in North America, and place it into a
museum compliant text format
}
