% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_parser.R
\name{date_parser}
\alias{date_parser}
\title{create herbarium format dates for specimens}
\usage{
date_parser(x, coll_date, det_date)
}
\arguments{
\item{x}{an (sf/tibble/) data frame with both the collection date
and determination date}

\item{coll_date}{date of collection, expected to be of the format 'MM/DD/YYYY', minor checks for compliance to the format will be carried out before returning an error.}

\item{det_date}{date of determination, same format and processes as above.}
}
\value{
original data frame plus: x_dmy, x_day, x_month, x_year, x_text, det_date_text, columns for both parameters which are supplied as inputs
}
\description{
this function will return up to two additional columns with x. 'coll_date_text' a text format for the date of collection and 'det_date_text" a shorter text format date for identification
}
