% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directions_grabber.R
\name{directions_grabber}
\alias{directions_grabber}
\title{Have google maps help you write directions to a site}
\usage{
directions_grabber(x, api_key)
}
\arguments{
\item{x}{an sf/tibble/dataframe of locations}

\item{api_key}{a google developer api key for use with googleway}
}
\description{
This function helps users write directions to a site which is somewhat near roads.
}
\examples{
# see package vignette
}
