% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms2dd.R
\name{dmsbyrow}
\alias{dmsbyrow}
\title{will operate row-wise in case of mixed coordinate system}
\usage{
dmsbyrow(x, long, lat)
}
\arguments{
\item{x}{an input data frame to apply operations to}

\item{long}{column holding longitude values}

\item{lat}{column holding latitude values.}
}
\description{
will operate row-wise in case of mixed coordinate system
}
\keyword{internal}
