% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{download_data}
\alias{download_data}
\title{Download the data required to establish a BL instance}
\usage{
download_data(path)
}
\arguments{
\item{path}{The root directory to save all the data in. Please specify a location, we suggest you make a directory for this.
If not specified will default to your working directory.}
}
\description{
this function will download the data which are required to set up an instance of BarnebyLives.
These data must then be processed by the \code{data_process} function to set up the directory structures appropriately.
Note that the function will test if the data already exist in the location, if they do they will not be downloaded again.
}
\examples{
\donttest{
# setwd('/media/steppe/hdd/BL_sandbox/geodata_raw')
# download_data(path = '/media/steppe/hdd/BL_sandbox')
}
}
