% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{download_sgmc}
\alias{download_sgmc}
\title{use to try and download sgmc data}
\usage{
download_sgmc(dest, url)
}
\arguments{
\item{dest}{location to save data to}

\item{url}{location to try download from}
}
\description{
use to try and download sgmc data
}
\keyword{internal}
