% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_spell_check.R
\name{family_spell_check}
\alias{family_spell_check}
\title{check that family is spelledcorrectly.}
\usage{
family_spell_check(x, path)
}
\arguments{
\item{x}{data frame/ tibble /sf object containing names to spell check}

\item{path}{a path to a folder containing the taxonomic data.}
}
\description{
this function attempts to verify the spelling of a user submitted taxonomic name.
}
\examples{
\dontrun{
names <- data.frame(
 Collection_number = 1:3,
 Family = c('Asteracea', 'Flabaceae', 'Onnagraceae')
)
spelling <- family_spell_check(names, path = '../taxonomic_data')
}
}
