% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_degree.R
\name{format_degree}
\alias{format_degree}
\title{ASCII compliant degree symbol}
\usage{
format_degree(x, symbol = "°")
}
\arguments{
\item{x}{a value to add the symbol onto.}

\item{symbol}{fixed here as the degree symbol}
}
\description{
This function is required to pass R CMD CHECKS via generating an appropriately
encoded degree symbol in line.
}
\examples{
format_degree(180)
}
\keyword{internal}
