% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{geological_map}
\alias{geological_map}
\title{Set up the downloaded data for a BarnebyLives instance}
\usage{
geological_map(path, pathOut, tile_cells)
}
\description{
used within \code{data_setup}
}
\keyword{internal}
