% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_directions.R
\name{get_google_directions}
\alias{get_google_directions}
\title{identify a populated place near a collection and get directions from there}
\usage{
get_google_directions(x, api_key)
}
\arguments{
\item{x}{an sf/tibble/dataframe of locations}

\item{api_key}{a google developer api key for use with googleway}
}
\description{
This function identifies the closest populated place, and makes a google API call
via the 'googleway' package to get directions from it to the collection site.
This function operates entirely within 'directions_grabber'
}
\examples{
# see 'directions_grabber'
}
