% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarnebyLives.r
\docType{data}
\name{google_towns}
\alias{google_towns}
\title{This is a reproduced and slightly curated list of populated places.}
\format{
Each is a tibble with two variables and three observations
}
\usage{
data(google_towns)
}
\description{
this sf/tibble/data.frame data set is derived from the U.S. Census Bureaus tigris::places() data set. It has only the centroid of each place rounded to
three decimal places to reduce size.
}
\keyword{datasets}
