% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herbaria_info.R
\docType{data}
\name{herbaria_info}
\alias{herbaria_info}
\title{herbaria_info}
\format{
Each is a data frame with columns relevant to letter heads
}
\usage{
herbaria_info
}
\description{
This is a small data frame of herbaria contact info
}
\details{
this data.frame data set is derived from Index Herbariorum and features some places many of our collectors send specimens to.
}
\examples{
head(herbaria_info)
}
\keyword{datasets}
