% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarnebyLives.r
\docType{data}
\name{ipni_authors}
\alias{ipni_authors}
\title{IPNI author abbreviations}
\format{
A character vector
}
\usage{
data(ipni_authors)
}
\description{
A vector from ipni of all standard author abbreviations
}
\keyword{datasets}
