% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_writer.R
\name{label_writer}
\alias{label_writer}
\title{write herbarium labels to pdf}
\usage{
label_writer(x, outdir)
}
\arguments{
\item{x}{a path to a csv of labels which have been cleaned with BL, and which the analyst
has evaluated for conflicts.}

\item{outdir}{an output directory to hold the files.}
}
\description{
This function uses the BarnebyLives custom template to write labels ala
a mail merge. Each label will be written as a separate file with the dimensions
of 4x4 inches.
}
\examples{
\dontrun{
label_writer(collection_examples, BL_label_example)
}
}
