% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{mason}
\alias{mason}
\title{a wrapper around terra::makeTiles for setting the domain of a project}
\usage{
mason(path, pathOut, tile_cellsV)
}
\arguments{
\item{path}{the input directory with all geographic data.}

\item{pathOut}{the output directory for the final data product to go.
Fn will automatically create folders for each product type.}

\item{tile_cellsV}{output of \code{make_tiles}}
}
\description{
for projects with large spatial domains or using relatively high resolution data, this will help

you make virtual tiles which do not need to be held in memory for the project.
}
\keyword{internal}
