% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{places_and_spaces}
\alias{places_and_spaces}
\title{Set up the downloaded data for a BarnebyLives instance}
\usage{
places_and_spaces(path, pathOut, bound)
}
\description{
used within \code{data_setup}
}
\keyword{internal}
