% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powNAce.R
\name{powNAce}
\alias{powNAce}
\title{Easily compare POW column to input columns}
\usage{
powNAce(x)
}
\arguments{
\item{x}{an sf/data frame/tibble which has both the input and POW ran data on it}
}
\description{
pronounced 'pounce'.
After querying the POW database via 'powo_searcher' this function will mark
POW cells which are identical to the input (verified cleaned) as NA
ideally making it easier for person to process there results and prepare
there data for sharing via labels and digitally.
}
\examples{
df <- data.frame(
 POW_Genus = c('Castilleja', 'Linnaea', 'Dimeresia'),
 POW_Epithet = c('pilosa', 'borealis', 'howellii'),
 POW_Infrarank = c('var.', 'var.', NA),
 POW_Infraspecies =  c('pilosa', 'americana', NA),
 POW_Authority =  c('(S. Watson) Rydb.', '(J. Forbes) Rehder', 'A. Gray')
)
powNAce(df)
}
