% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{process_geology}
\alias{process_geology}
\title{Set up the downloaded data for a BarnebyLives instance}
\usage{
process_geology(path, pathOut, tile_cells)
}
\description{
used within \code{data_setup}
}
\keyword{internal}
