% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{process_gnis}
\alias{process_gnis}
\title{Set up the downloaded data for a BarnebyLives instance}
\usage{
process_gnis(path, pathOut, bound)
}
\description{
used within \code{data_setup}
}
\keyword{internal}
