% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_examples.R
\docType{data}
\name{project_examples}
\alias{project_examples}
\title{project_examples}
\format{
A data frame with projects and contact info
}
\usage{
project_examples
}
\description{
This is a small data frame of examples of collection projects
}
\details{
This is a small data frame of examples of collection projects
}
\examples{
head(project_examples)
}
\keyword{datasets}
