% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarnebyLives.r
\docType{data}
\name{randomNames}
\alias{randomNames}
\alias{names_vec}
\title{10 random plant names from taxize}
\format{
A character vector
}
\usage{
data(names_vec)
}
\description{
A vector from taxize::names_list(rank = 'species', size = 10)
}
\keyword{datasets}
