% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_grabber.R
\name{result_grabber}
\alias{result_grabber}
\title{collect the results of a successful powo search}
\usage{
result_grabber(x)
}
\arguments{
\item{x}{a successful powo search query}
}
\description{
this function is used within 'powo_searcher' to retrieve the results from a successful POWO query.
}
