% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipping_examples.R
\docType{data}
\name{shipping_examples}
\alias{shipping_examples}
\title{shipping_examples}
\format{
A data frame with some collections which will one day be shipped.
}
\usage{
shipping_examples
}
\description{
This is a small data frame of species information for shipping manifests
}
\details{
this data frame is derived from collections in 2023.
}
\examples{
head(shipping_examples)
}
\keyword{datasets}
