% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_writer.R
\name{site_writer}
\alias{site_writer}
\title{gather distance and azimuth from shop to center of town}
\usage{
site_writer(x, path)
}
\arguments{
\item{x}{an sf/tibble/dataframe of locations with associated nearest locality data}

\item{path}{path to gnis_places}
}
\description{
Help provide some simple context between the building and the
}
\examples{
\dontrun{
library(sf)
sites <- data.frame(
 longitude_dd = runif(15, min = -120, max = -100),
 latitude_dd = runif(15, min = 35, max = 48)
 ) |>
 sf::st_as_sf(coords = c('longitude_dd', 'latitude_dd'), crs = 4326)

head(sites)
distaze_results <- distAZE(sites, path = '/hdd/Barneby_Lives-dev/geodata/places') # takes some time
head(distaze_results)
}
}
