% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_font.R
\name{species_font}
\alias{species_font}
\title{ensure proper italicization of a associated species}
\usage{
species_font(x)
}
\arguments{
\item{x}{a value being feed into the label during the merge}
}
\description{
This function works at the time of label creation to ensure that abbreviations
such as 'species' 'variety' and such are not italicized, per convention.
}
\examples{
associates <- c('Eriogonum ovalifolium var. purpureum, Castilleja sp., Crepis spp.')
species_font(associates)
}
