% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificDirections.R
\name{specificDirections}
\alias{specificDirections}
\title{Import Google directions to a site}
\usage{
specificDirections(x)
}
\arguments{
\item{x}{the output of directions_grabber}
}
\description{
Use Google Maps to write directions to a site, so that the people re-locating
the site don't need to use Google Maps to write directions to a site. These
are crude and will require some refinement for some relevant applications.
This function operates entirely within 'directions_grabber'
}
\examples{
# see 'directions_grabber'#
}
