% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_scientificName.R
\name{split_scientificName}
\alias{split_scientificName}
\title{split out a scientific input column to pieces}
\usage{
split_scientificName(x, sciName_col, overwrite)
}
\arguments{
\item{x}{Dataframe with collection information.}

\item{sciName_col}{Character. Column containing the data to parse}

\item{overwrite}{Boolean. Whether to overwrite the original input columns,
or simply append the spell checked columns.}
}
\description{
This function attempts to split a scientific name into it's component pieces.
Given an input scientific, or scientific with scientific authorities and infraspecies
this function will parse them into the columns used in the BarnebyLives pipeline.
}
\examples{
library(BarnebyLives)
ce <- collection_examples
ce <- collection_examples[
 sample(1:nrow(collection_examples), 25), 
 Sci_name = c('Collection_number', 'Name_authority')
]

split_scientificName(ce, sciName_col = 'Sci_name')|> head()
}
