% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxUnpack.R
\name{TaxUnpack}
\alias{TaxUnpack}
\title{a function to subset the WCVP data set to an area of focus}
\usage{
TaxUnpack(path, bound)
}
\arguments{
\item{path}{Character vector. Location to where taxonomic data should be saved,
we recommend a subdirectory in the same folder, and at the same level, as the geographic data.}

\item{bound}{Dataframe of x and y coordinates, same argument as to \code{data_setup}.
Coordinates assumed in WG84, NAD83, or googles absurd projection. Coordinate
specifications are very precise later in the process, but at this level any of these
three are effectively equivalent.}
}
\description{
This function is run on the WCVP compressed archive downloaded by 'wcvp_update',
and requires some input from the user which specifies a geographic area to establish a
taxonomic look up table for. Writes four lookup tables: a species catalog, an
infraspecific taxa catalog (varieties and subspecies), a family catalog, and a
global (unfiltered by geography) taxon ID index used by \code{\link{wcvp_searcher}}
to resolve synonyms to their currently accepted name.
}
\examples{
\dontrun{

bound <- bound |>
  sf::st_as_sf(coords = c('x', 'y'), crs = 4326) |>
  sf::st_bbox() |>
  sf::st_as_sfc()

oupu <- TaxUnpack(path = '/media/steppe/hdd/BL_sandbox/taxdata-MI', bound = bound)
}
}
