% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_again.R
\name{try_again}
\alias{try_again}
\title{try to download from Kew again, with just the binomial}
\usage{
try_again(x)
}
\arguments{
\item{x}{output from the first step of 'powo_searcher'}
}
\description{
several infra species fail from POWO, retry the query with just the base name. This not accessed directly by the user but used inside 'powo_searcher'
}
