% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USDA_plants_lkp.R
\name{USDA_plants_lkp}
\alias{USDA_plants_lkp}
\title{Retrieve binomials from a list of USDA Plants codes}
\usage{
USDA_plants_lkp(x, y, path)
}
\arguments{
\item{x}{a data frame containing the values}

\item{y}{the first column containing the codes, the function works row by row allowing
for mixed columns, i.e. those with both spelled out and abbreviated species}

\item{path}{path to the folder containing taxonomic data}
}
\description{
Spell out abbreviated fields
}
