% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcvp_update.R
\name{wcvp_update}
\alias{wcvp_update}
\title{install or update World Vascular Plants Checklist}
\usage{
wcvp_update(tax_dat_p)
}
\arguments{
\item{verbose}{boolean, TRUE/FALSE, whether to print messages to console or not. Defaults to FALSE.}

\item{tax_data_path}{the path to the taxonomic data, the raw wcvp.zip should be there.}
}
\description{
This function checks to see whether the version of WCVP is the most current, if not, it will re-download WCVP and set up the BL taxonomy structure.
}
\examples{
\dontrun{
wcvp_update('/home/sagesteppe/Downloads')
}
}
