% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{writer}
\alias{writer}
\title{write values and collapse NAs}
\usage{
writer(x, italics)
}
\arguments{
\item{x}{the input character of length 1}

\item{italics}{italicize or not? Boolean, defualts to FALSE}
}
\description{
This function will determine whether to print or hide a variable onto labels
}
\examples{
writer(collection_examples[9,'Binomial_authority'] )
writer(collection_examples[9, 'Infraspecies'], italics = TRUE )
}
