% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer2.R
\name{writer2}
\alias{writer2}
\title{write values and collapse NAs}
\usage{
writer2(x, italics, period)
}
\arguments{
\item{x}{the input character of length 1}

\item{italics}{italicize or not? Boolean, defaults to FALSE}

\item{period}{Boolean, whether to add a period or not at end of string.
Defaults to FALSE.}
}
\description{
This function will determine whether to print a variable onto labels or hide it
}
\examples{
writer(collection_examples[9,'Binomial_authority'] )
writer(collection_examples[9, 'Infraspecies'], italics = TRUE )
}
