% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{writer_fide}
\alias{writer_fide}
\title{do or don't write determination information}
\usage{
writer_fide(x)
}
\arguments{
\item{x}{data frame holding the values}
}
\description{
This functions determines whether to write information on whom determined the specimen
when, and via which texts. It used as a final step when formatting labels.
}
\examples{
text <- data.frame(
Fide = c('Michigan Flora Book (emphatic)',
        'Michigan Flora Book (emphatic) verbose',
        'Michigan Flora Book'),
Determined_by = rep('Marabeth', 3),
Determined_date_text = rep('a day in time', 3)
) |>
 split(f = 1:3)

lapply(text, writer_fide)
}
