% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate_dropper.R
\name{associate_dropper}
\alias{associate_dropper}
\title{remove collection from associated species}
\usage{
associate_dropper(x, Binomial, col)
}
\arguments{
\item{x}{data frame containing collection info}

\item{Binomial}{Character. a column containing the name of the collection, without authorship.
Defaults to 'Full_name'.}

\item{col}{Character. the column with the species to be checked}
}
\description{
Removes the collected species from the list of associated species
}
\examples{
ce <- data.frame(
Full_name = c(
  'Microseris gracilis', 'Alyssum desertorum', 'Ceratocephala testiculata'),
Associates = rep(
   'Microseris gracilis, Lupinus lepidus, Alyssum desertorum, Ceratocephala testiculata',
 times = 3)
)

associate_dropper(ce, col = 'Associates')
}
