% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_examples.R
\docType{data}
\name{collection_examples}
\alias{collection_examples}
\title{collection_examples}
\format{
A data frame with some collections which will one day be shipped.
}
\usage{
collection_examples
}
\description{
This is a small data frame of species information for use with the shipping/transferral notes
as well as with the label makers.
}
\details{
this data frame is derived from collections in 2023, which have been cleaned with the
BarnebyLives pipeline
}
\examples{
head(collection_examples)
}
\keyword{datasets}
