% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directions_overview.R
\name{directions_overview}
\alias{directions_overview}
\title{return a quick route summary from googleway::google_directions}
\usage{
directions_overview(x)
}
\arguments{
\item{x}{results from a googleway::google_directions API query}
}
\description{
This function creates a quick abstract of a google maps route, which contains the travel time, starting location and the route overview. This function operates entirely within 'directions_grabber'.
}
\examples{
# see 'directions_grabber'
}
