% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_towns.R
\docType{data}
\name{google_towns}
\alias{google_towns}
\title{google_towns}
\format{
Each is a tibble with two variables and three observations
}
\usage{
google_towns
}
\description{
This is a reproduced and slightly curated list of populated places.
}
\details{
this sf/tibble/data.frame data set is derived from the U.S. Census Bureaus tigris::places() data set. It has only the centroid of each place rounded to
three decimal places to reduce size.
}
\examples{
head(google_towns)
}
\keyword{datasets}
