% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_maker.R
\name{map_maker}
\alias{map_maker}
\title{make a quick county dot map to display the location of the collection}
\usage{
map_maker(x, states, path, collection_col)
}
\arguments{
\item{x}{an sf dataframe of coordinates to make maps for, requires collection number and spatial attributes}

\item{path}{a directory to store the map images in before merging}

\item{collection_col}{column specify the collection number or other UNIQUE id for the collection}
}
\description{
use sf to create a 20th centtury style 'dot map' which features the state
boundary and county lines.
}
