% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mason.R
\name{mason}
\alias{mason}
\title{a wrapper around terra::makeTiles for setting the domain of a project}
\usage{
mason(dirin, dirout, grid, fnames, ...)
}
\arguments{
\item{dirin}{the input directory, including path, with the dataset}

\item{dirout}{the output directory, including path, for the subset dataset to go}

\item{grid}{spatvector defining the tiles, should contain a column containing grid cell names}

\item{fnames}{name of column containing filenames in the grid}

\item{...}{other arguments passed to file.path}
}
\description{
for projects with large spatial domains or using relatively high resolution data, this will help

you make virtual tiles which do not need to be held in memory for the project.
}
\examples{
# see package vignette
}
