% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notFound.R
\name{notFound}
\alias{notFound}
\title{notify user if an entry had any results not found in POWO}
\usage{
notFound(x)
}
\arguments{
\item{x}{output of 'powo_searcher' after binding rows}

\item{returns}{messages to consoles indicating search terms, and there status if failed to be found. This desirable because 'powo_searcher' squashes these errors.}
}
\description{
simple function to run on 'powo_searcher' results to show species not found which
}
