% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notFound.R
\name{notFound}
\alias{notFound}
\title{notify user if an entry had any results not found in POWO}
\usage{
notFound(x)
}
\arguments{
\item{x}{output of 'powo_searcher' after binding rows}
}
\value{
messages to consoles indicating search terms, and there status if failed to be found. This desirable because 'powo_searcher' squashes these errors.
}
\description{
simple function to run on 'powo_searcher' results to show species not found which
}
\examples{
library(dplyr)
library(crayon)
names_vec <- taxize::names_list(rank = 'species', size = 10)
# 10 random species from taxize, usually 1 or 2 species are not found in Plants of the world online
pow_results <- lapply(names_vec, powo_searcher) |>
  dplyr::bind_rows()
# pow_results[,1:5]
# if there is not a family which is 'NOT FOUND', reshuffle the random species from taxize.
notFound(pow_results) # little message.
}
