% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physical_grabber.R
\name{physical_grabber}
\alias{physical_grabber}
\title{gather physical characteristics of the site}
\usage{
physical_grabber(x)
}
\arguments{
\item{x}{an sf data frame of collection points}

\item{y}{a column which unambiguously identifies each collection}
}
\description{
this function grabs information on the elevation, azimuth, geomorphon, and geology of the site
}
\examples{
see the package vignette
}
