% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places.R
\docType{data}
\name{places}
\alias{places}
\title{Places}
\format{
Each is a tibble with two variables and three observations
}
\usage{
places
}
\description{
This is a reproduced and slighly curated list of populated places.
}
\details{
this sf/tibble/data.frame data set is derived from the U.S. Census Bureaus tigriss::places() dataset. It has only the centroid of each place rounded to 
three decimal places to reduce size.
}
\examples{
head(places)
}
\keyword{datasets}
