% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/political_grabber.R
\name{political_grabber}
\alias{political_grabber}
\title{gather political site information}
\usage{
political_grabber(x, y)
}
\arguments{
\item{x}{an sf data frame of collection points}

\item{y}{a column which unambiguously identifies each collection}
}
\description{
this function grabs information on the state, county, and township of collections
}
\examples{
see the package vignette
}
