% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/political_grabber.R
\name{political_grabber}
\alias{political_grabber}
\title{gather political site information}
\usage{
political_grabber(x, y, path)
}
\arguments{
\item{x}{an sf data frame of collection points}

\item{y}{a variable which uniquely identifies each observation}

\item{path}{a path to the directory holding the BarnebyLivesGeodata}
}
\description{
this function grabs information on the state, county, and township of collections
}
\examples{
# see the package vignette
}
