% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powo_searcher.R
\name{powo_searcher}
\alias{powo_searcher}
\title{query plants of the world online for taxonomic information}
\usage{
powo_searcher(x)
}
\arguments{
\item{x}{a vector of species names to submit, these should have clean spelling
notes: results are observed to fail for valid infraspecies on Kew's end, and they seem not
to mention valid infraspecies.}
}
\description{
a wrapper for kewr::search_powo
}
