% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipping_examples.R
\docType{data}
\name{shipping_examples}
\alias{shipping_examples}
\title{shipping_examples}
\format{
data frame
}
\usage{
shipping_examples
}
\description{
A list of a set of some specimens the package author once shipped.
}
\details{
this data.frame is based on some specimens I sent somewhere, or it's randomly sampled from collection.
}
\examples{
head(shipping_examples)
}
\keyword{datasets}
