% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell_check.R
\name{spell_check}
\alias{spell_check}
\title{check that genera and specific epithets are spelled (almost) correctly}
\usage{
spell_check(x)
}
\arguments{
\item{x}{a vector of species names}
}
\description{
this function attempts to verify the spelling of a user submitted taxonomic name. If necessary it will proceed step-wise by name pieces attempting to place them.
}
