% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell_check.R
\name{spell_check}
\alias{spell_check}
\title{check that genera and specific epithets are spelled (almost) correctly}
\usage{
spell_check(x, path)
}
\arguments{
\item{x}{data frame/ tibble /sf object containing names to spell check}

\item{path}{a path to a folder containing the taxonomic data.}

\item{y}{a column in it containing a full scientific name}
}
\description{
this function attempts to verify the spelling of a user submitted taxonomic name. If necessary it will proceed step-wise by name pieces attempting to place them.
}
\examples{
\dontrun{
names_vec <- c('Astagalus purshii', 'Linnaeus borealius', 'Heliumorus multifora')
spelling <- spell_check(names_vec, path = '../taxonomic_data')
spelling
}
}
