% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxUnpack.R
\name{TaxUnpack}
\alias{TaxUnpack}
\title{a function to subset the WCVP data set to an area of focus}
\usage{
TaxUnpack(path, continents)
}
\arguments{
\item{path}{to where WCVP is stored}

\item{continents}{one of: 'NORTHERN AMERICA', 'AFRICA', 'ANTARCTICA', 'ASIA-TROPICAL',
'ASIA-TEMPERATE', 'AUSTRALASIA', 'EUROPE', 'OCEANIA', 'SOUTHERN AMERICA', 'PACIFIC',}
}
\description{
This function is run on the WCVP compressed archive downloaded by 'wcvp_update',
and requires some input from the user which specifies a geographic area to establish a
taxonomic look up table for.
}
\examples{
\dontrun{
oupu <- TaxUnpack(path = '~/Downloads',
                 continent = 'NORTHERN AMERICA')

}
}
