% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer_fide.R
\name{writer_fide}
\alias{writer_fide}
\title{do or don't write determination information}
\usage{
writer_fide(x)
}
\arguments{
\item{x}{data frame holding the values}
}
\description{
This functions determines whether to write information on whom determined the specimen
when, and via which texts. It used as a final step when formatting labels.
}
\examples{
phacelia <- collection_examples[31,]
writer_fide(phacelia)
}
